% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safeSave-package.R
\docType{package}
\name{safeSave-package}
\alias{safeSave}
\alias{safeSave-package}
\title{safeSave: Find a File or Folder Name That Does Not Already Exist}
\description{
Small tool consisting of a single function that helps you find a 
    file or folder name that does not already exist
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cisglee/safeSave}
  \item Report bugs at \url{https://github.com/cisglee/safeSave/issues}
}

}
\keyword{internal}
